open_midas/requirements.txt
torch==1.13.1
torchvision==0.14.1
imgaug==0.4.0
matplotlib==3.7.1
numpy==1.22.3
seaborn==0.12.2
IPython==8.7.0
Pillow==9.3.0
tqdm==4.64.1
scikit-learn==1.2.0
networkx==2.8.4
opencv-python==4.7.0.68
catboost==1.2